Renamed main slice to table slice
