+++ src/redux/hooks.ts
@@ -6,8 +6,8 @@ import { AppDispatch, RootState } from "./store";
 export const useAppDispatch = () => useDispatch<AppDispatch>();
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
 
-export const useCurrentDate: () => string = () => useAppSelector(state => state.main.currentDate);
-export const useStartDate: () => string = () => useAppSelector(state => state.main.startDate);
+export const useCurrentDate: () => string = () => useAppSelector(state => state.table.currentDate);
+export const useStartDate: () => string = () => useAppSelector(state => state.table.startDate);
 export const useColumns: () => number = () => useAppSelector(state => state.columns.value);
 export const useGrabbedTile: () => GrabbedTileState = () => useAppSelector(state => state.grabbedTile);
 export const useHotel: () => HotelData = () => useAppSelector(state => state.hotel.data);
