+++ src/redux/root.ts
@@ -1,9 +1,9 @@
 import { combineReducers } from "@reduxjs/toolkit";
 
-import main from "./mainSlice";
+import table from "./tableSlice";
 import grabbedTile from "./grabbedTileSlice";
 import hotel from "./hotelSlice";
 import scroll from "./scrollSlice";
 import columns from "./columnsSlice";
 
-export const reducer = combineReducers({ main, grabbedTile, hotel, scroll, columns });
+export const reducer = combineReducers({ table, grabbedTile, hotel, scroll, columns });
